Fix inconsistent casting in logistic_regression.cpp idx_col calls

+++ src/logistic_regression.cpp
@@ -53,7 +53,7 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
   for (size_t col = 0; col < pp; ++col) {
     double beta = par[col];
     if (beta == 0.0) continue;
-    size_t off = FlatMatrix::idx_col(0, static_cast<int>(col), n);
+    size_t off = FlatMatrix::idx_col(0, static_cast<std::size_t>(col), n);
     for (size_t r = 0; r < nn; ++r) {
       eta[r] += beta * Z.data[off + r];
     }
@@ -108,7 +108,7 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     
     // accumulate score: score[i] += f*w*v*score_scale*z(person,i)
     for (size_t i = 0; i < pp; ++i) {
-      size_t zoff = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+      size_t zoff = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
       double zi = Z.data[zoff + person];
       score[i] += f * w * v * score_scale * zi;
     }
@@ -116,12 +116,12 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     // accumulate information matrix (lower triangle)
     // imat[i,j] += f*w*v_for_info*z_i*z_j  for j <= i (we'll mirror later)
     for (size_t i = 0; i < pp; ++i) {
-      size_t off_i = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+      size_t off_i = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
       double zi = Z.data[off_i + person];
       for (size_t j = 0; j <= i; ++j) {
-        size_t off_j = FlatMatrix::idx_col(0, static_cast<int>(j), n);
+        size_t off_j = FlatMatrix::idx_col(0, static_cast<std::size_t>(j), n);
         double zj = Z.data[off_j + person];
-        size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p); // row=i, col=j in p x p
+        size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p); // row=i, col=j in p x p
         imat.data[idx] += f * w * v_for_info * zi * zj;
       }
     }
@@ -153,8 +153,8 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
   // mirror lower to upper triangle of imat
   for (size_t i = 0; i < pp; ++i) {
     for (size_t j = i + 1; j < pp; ++j) {
-      size_t idx_lower = FlatMatrix::idx_col(static_cast<std::size_t>(j), static_cast<int>(i), p); // stored at row=j, col=i
-      size_t idx_upper = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p); // row=i, col=j
+      size_t idx_lower = FlatMatrix::idx_col(static_cast<std::size_t>(j), static_cast<std::size_t>(i), p); // stored at row=j, col=i
+      size_t idx_upper = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p); // row=i, col=j
       imat.data[idx_upper] = imat.data[idx_lower];
     }
   }
@@ -165,7 +165,7 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     cholesky2(imat0, p, 1e-12);
     double vdet = 0.0;
     for (size_t i = 0; i < pp; ++i) {
-      size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(i), p);
+      size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(i), p);
       vdet += std::log(imat0.data[idx]);
     }
     double penloglik = loglik + 0.5 * vdet;
@@ -177,12 +177,12 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
       double w = param->weight.empty() ? 1.0 : param->weight[person];
       double mult = f * w * a[person];
       for (size_t i = 0; i < pp; ++i) {
-        size_t off_i = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+        size_t off_i = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
         double zi = Z.data[off_i + person];
         for (size_t j = 0; j <= i; ++j) {
-          size_t off_j = FlatMatrix::idx_col(0, static_cast<int>(j), n);
+          size_t off_j = FlatMatrix::idx_col(0, static_cast<std::size_t>(j), n);
           double zj = Z.data[off_j + person];
-          size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p);
+          size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p);
           xwx.data[idx] += mult * zi * zj;
         }
       }
@@ -190,8 +190,8 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     // mirror
     for (size_t i = 0; i < pp; ++i) {
       for (size_t j = i + 1; j < pp; ++j) {
-        size_t idx_lower = FlatMatrix::idx_col(static_cast<std::size_t>(j), static_cast<int>(i), p);
-        size_t idx_upper = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p);
+        size_t idx_lower = FlatMatrix::idx_col(static_cast<std::size_t>(j), static_cast<std::size_t>(i), p);
+        size_t idx_upper = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p);
         xwx.data[idx_upper] = xwx.data[idx_lower];
       }
     }
@@ -208,18 +208,18 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
       // compute h = z^T * var * z  (z is column vector length p for this person)
       double h = 0.0;
       for (size_t i = 0; i < pp; ++i) {
-        size_t off_i = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+        size_t off_i = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
         double zi = Z.data[off_i + person];
         for (size_t j = 0; j < pp; ++j) {
-          size_t idx_var = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p);
-          h += varf.data[idx_var] * zi * Z.data[FlatMatrix::idx_col(0, static_cast<int>(j), n) + person];
+          size_t idx_var = FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p);
+          h += varf.data[idx_var] * zi * Z.data[FlatMatrix::idx_col(0, static_cast<std::size_t>(j), n) + person];
         }
       }
       h *= mult;
       double resid = param->y[person] - pi[person];
       double u = f * w * resid * d[person] + 0.5 * b[person] * h;
       for (size_t i = 0; i < pp; ++i) {
-        size_t off_i = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+        size_t off_i = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
         double zi = Z.data[off_i + person];
         g[i] += u * zi;
       }
@@ -232,7 +232,7 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     std::vector<std::vector<double>> imat_out(pp, std::vector<double>(pp));
     for (size_t i = 0; i < pp; ++i)
       for (size_t j = 0; j < pp; ++j)
-        imat_out[i][j] = imat.data[FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p)];
+        imat_out[i][j] = imat.data[FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p)];
     result.push_back(imat_out, "imat");
     result.push_back(loglik, "regloglik");
     result.push_back(score, "regscore");
@@ -243,7 +243,7 @@ ListCpp f_der_0(int p, const std::vector<double>& par, void *ex, bool firth) {
     std::vector<std::vector<double>> imat_out(pp, std::vector<double>(pp));
     for (size_t i = 0; i < pp; ++i)
       for (size_t j = 0; j < pp; ++j)
-        imat_out[i][j] = imat.data[FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<int>(j), p)];
+        imat_out[i][j] = imat.data[FlatMatrix::idx_col(static_cast<std::size_t>(i), static_cast<std::size_t>(j), p)];
     result.push_back(loglik, "loglik");
     result.push_back(score, "score");
     result.push_back(imat_out, "imat");
@@ -271,7 +271,7 @@ FlatMatrix f_ressco_0(int p, const std::vector<double>& par, void *ex) {
   for (size_t col = 0; col < pp; ++col) {
     double beta = par[col];
     if (beta == 0.0) continue;
-    size_t off = FlatMatrix::idx_col(0, static_cast<int>(col), n);
+    size_t off = FlatMatrix::idx_col(0, static_cast<std::size_t>(col), n);
     for (size_t r = 0; r < nn; ++r) {
       eta[r] += beta * Z.data[off + r];
     }
@@ -299,9 +299,9 @@ FlatMatrix f_ressco_0(int p, const std::vector<double>& par, void *ex) {
     }
     double v = param->y[person] - r;
     for (size_t i = 0; i < pp; ++i) {
-      size_t zoff = FlatMatrix::idx_col(0, static_cast<int>(i), n);
+      size_t zoff = FlatMatrix::idx_col(0, static_cast<std::size_t>(i), n);
       double zi = Z.data[zoff + person];
-      size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(person), static_cast<int>(i), n);
+      size_t idx = FlatMatrix::idx_col(static_cast<std::size_t>(person), static_cast<std::size_t>(i), n);
       resid.data[idx] = v * dscale * zi;
     }
   }
